[Java] Extract common base proxy class.
