+++ aeron-driver/src/main/java/io/aeron/driver/DriverConductorProxy.java
@@ -19,32 +19,22 @@ import io.aeron.driver.media.ReceiveChannelEndpoint;
 import io.aeron.driver.media.ReceiveDestinationTransport;
 import io.aeron.driver.media.SendChannelEndpoint;
 import io.aeron.driver.media.UdpChannel;
-import org.agrona.concurrent.AgentTerminationException;
 import org.agrona.concurrent.QueuedPipe;
 import org.agrona.concurrent.status.AtomicCounter;
 
 import java.net.InetSocketAddress;
 
-import static io.aeron.driver.ThreadingMode.INVOKER;
-import static io.aeron.driver.ThreadingMode.SHARED;
-
 /**
  * Proxy for sending commands to the {@link DriverConductor}.
  */
-public final class DriverConductorProxy
+public final class DriverConductorProxy extends CommandProxy
 {
-    private final ThreadingMode threadingMode;
-    private final QueuedPipe<Runnable> commandQueue;
-    private final AtomicCounter failCount;
-
     private DriverConductor driverConductor;
 
     DriverConductorProxy(
         final ThreadingMode threadingMode, final QueuedPipe<Runnable> commandQueue, final AtomicCounter failCount)
     {
-        this.threadingMode = threadingMode;
-        this.commandQueue = commandQueue;
-        this.failCount = failCount;
+        super(threadingMode, commandQueue, failCount);
     }
 
     /**
@@ -161,36 +151,6 @@ public final class DriverConductorProxy
         }
     }
 
-    /**
-     * Is the driver conductor not concurrent with the sender and receiver threads.
-     *
-     * @return true if the {@link DriverConductor} is on the same thread as the sender and receiver.
-     */
-    public boolean notConcurrent()
-    {
-        return threadingMode == SHARED || threadingMode == INVOKER;
-    }
-
-    /**
-     * Get the threading mode of the driver.
-     * @return ThreadingMode of the driver.
-     */
-    public ThreadingMode threadingMode()
-    {
-        return threadingMode;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public String toString()
-    {
-        return "DriverConductorProxy{" +
-            "threadingMode=" + threadingMode +
-            ", failCount=" + failCount +
-            '}';
-    }
-
     void driverConductor(final DriverConductor driverConductor)
     {
         this.driverConductor = driverConductor;
@@ -244,20 +204,4 @@ public final class DriverConductorProxy
         }
     }
 
-    private void offer(final Runnable cmd)
-    {
-        while (!commandQueue.offer(cmd))
-        {
-            if (!failCount.isClosed())
-            {
-                failCount.increment();
-            }
-
-            Thread.yield();
-            if (Thread.currentThread().isInterrupted())
-            {
-                throw new AgentTerminationException("interrupted");
-            }
-        }
-    }
 }
